+++ modules/farm_rothamsted_experiment/src/Plugin/views/filter/ColumnLevel.php
@@ -8,7 +8,7 @@ use Drupal\views\Plugin\views\filter\InOperator;
 /**
  * Custom views filter for the plot column levels.
  *
- * Reads column levels from the plan specified by the views contextual filter.
+ * @see farm_rothamsted_experiment_views_pre_view
  *
  * @ViewsFilter("column_level")
  */
@@ -24,43 +24,10 @@ class ColumnLevel extends InOperator {
       return $this->valueOptions;
     }
 
-    // Check if a plan id is provided.
-    // @todo Should the plan argument or id be configuration for the filter?
-    if (isset($this->view->args[0])) {
-
-      // Load the plan.
-      $plan_id = $this->view->args[0];
-      $plan = Plan::load($plan_id);
-      if (!empty($plan) && $plan->hasField('column_descriptors') && !$plan->get('column_descriptors')->isEmpty()) {
-
-        // Load columns from json.
-        $columns = json_decode($plan->get('column_descriptors')->value);
-
-        // Build options for each column.
-        $column_options = [];
-        foreach ($columns as $column) {
-
-          // Build options for each column level.
-          // Use a comma to separate each option as "column_id,level_id"
-          // since commas are unlikely to be used in either ID.
-          $column_level_options = [];
-          foreach ($column->column_levels as $column_level) {
-            $value = $column->column_id . ',' . $column_level->level_id;
-            $column_level_options[$value] = "$column_level->level_name ($column_level->level_id)";
-          }
-
-          // Build label for the column.
-          $type_label = "$column->column_name ($column->column_id)";
-          $column_options[$type_label] = $column_level_options;
-        }
-
-        // Set the value options.
-        $this->valueOptions = $column_options;
-      }
-    }
-    // If no plan is provided, default to no options.
-    else {
-      $this->valueOptions = [];
+    // Populate value options from filter settings.
+    $this->valueOptions = [];
+    if ($options = $this->options['settings']['column_options']) {
+      $this->valueOptions = $options;
     }
 
     return $this->valueOptions;
@@ -73,19 +40,24 @@ class ColumnLevel extends InOperator {
     if (empty($this->value)) {
       return;
     }
-    $this->ensureMyTable();
 
-    // Create an additional OR group.
-    $group = count($this->query->where) + 1;
-    $this->query->setWhereGroup('OR', $group);
+    // Build a custom join instead of using $this->ensureMyTable().
+    // This lets us do a left outer join on the column_descriptor_key column.
+    // This is much more efficient if multiple column descriptors are actively
+    // being filtered.
+    $column_id = $this->options['settings']['column_id'];
+    $join = $this->getJoin();
+    $join->extra = [
+      [
+        'field' => 'column_descriptors_key',
+        'value' => $column_id,
+      ],
+    ];
+    $this->tableAlias = $this->query->addTable($this->table, $this->relationship, $join);
 
-    // Add a where expression for each factor key and value.
-    foreach ($this->value as $value) {
-      $values = explode(',', $value);
-      $key_placeholder = $this->placeholder();
-      $value_placeholder = $this->placeholder();
-      $this->query->addWhereExpression($group, "$this->tableAlias.column_descriptors_key = $key_placeholder AND $this->tableAlias.column_descriptors_value = $value_placeholder", [$key_placeholder => $values[0], $value_placeholder => $values[1]]);
-    }
+    // Add a where expression for the column descriptor value.
+    $value_placeholder = $this->placeholder() . '[]';
+    $this->query->addWhereExpression(NULL, "$this->tableAlias.column_descriptors_value IN($value_placeholder)", [$value_placeholder => $this->value]);
   }
 
 }